extract `UpgradeDependencies` build rewrite from command
