+++ bleep-cli/src/scala/bleep/commands/BuildUpdateDeps.scala
@@ -1,10 +1,10 @@
 package bleep
 package commands
 
-import bleep.commands.BuildUpdateDeps.ContextualDep
 import bleep.internal.CoursierLogger
 import bleep.logging.Logger
-import bleep.rewrites.{normalizeBuild, BuildRewrite}
+import bleep.model.{CrossProjectName, Dep}
+import bleep.rewrites.{normalizeBuild, UpgradeDependencies}
 import coursier.Repository
 import coursier.cache.FileCache
 import coursier.core.{Dependency, Versions}
@@ -21,26 +21,26 @@ case class BuildUpdateDeps(started: Started) extends BleepCommand {
     // a bleep dependency may be instantiated into several different coursier dependencies
     // depending on which scala versions and platforms are plugging in
     // collect all instantiations into this structure
-    val allDeps: Map[ContextualDep, Dependency] =
+    val allDeps: Map[UpgradeDependencies.ContextualDep, Dependency] =
       BuildUpdateDeps.instantiateAllDependencies(build)
 
     val config = started.lazyConfig.forceGet
     val repos = CoursierResolver.coursierRepos(build.resolvers.values, config.authentications).filter(_.repr.contains("http"))
     val fileCache = FileCache[Task]().withLogger(new CoursierLogger(started.logger))
 
-    val foundByDep: Map[ContextualDep, (Dependency, Versions)] = {
+    val foundByDep: Map[UpgradeDependencies.ContextualDep, (Dependency, Versions)] = {
       implicit val ec: ExecutionContext = started.executionContext
       Await.result(BuildUpdateDeps.fetchAllVersions(fileCache, repos, allDeps), Duration.Inf)
     }
 
-    val upgrades: Map[ContextualDep, model.Dep] =
+    val upgrades: Map[UpgradeDependencies.ContextualDep, model.Dep] =
       foundByDep.flatMap { case (tuple @ (bleepDep, _), (_, version)) =>
         val latest = version.latest
         if (latest == bleepDep.version) None
         else Some(tuple -> bleepDep.withVersion(latest))
       }
 
-    val newBuild = BuildUpdateDeps.upgradedBuild(started.logger, upgrades, build)
+    val newBuild = UpgradeDependencies(new BuildUpdateDeps.UpgradeLogger(started.logger), upgrades)(build)
     val newBuild1 = normalizeBuild(newBuild)
     yaml.writeShortened(newBuild1.file, started.buildPaths.bleepYamlFile)
 
@@ -49,9 +49,14 @@ case class BuildUpdateDeps(started: Started) extends BleepCommand {
 }
 
 object BuildUpdateDeps {
-  type ContextualDep = (model.Dep, model.VersionCombo)
+  class UpgradeLogger(logger: Logger) extends UpgradeDependencies.UpgradeLogger {
+    override def upgraded(project: CrossProjectName, dep: Dep, newVersion: String): Unit =
+      logger
+        .withContext(project)
+        .info(s"${dep.organization.value}:${dep.baseModuleName.value} ${dep.version} => $newVersion")
+  }
 
-  def instantiateAllDependencies(build: model.Build): Map[ContextualDep, Dependency] =
+  def instantiateAllDependencies(build: model.Build): Map[UpgradeDependencies.ContextualDep, Dependency] =
     build.explodedProjects
       .flatMap { case (crossName, p) =>
         val versionCombo = model.VersionCombo.fromExplodedProject(p).orThrowTextWithContext(crossName)
@@ -59,25 +64,6 @@ object BuildUpdateDeps {
           case dep if !dep.version.contains("$") => ((dep, versionCombo), dep.asDependency(versionCombo).orThrowTextWithContext(crossName))
         }
       }
-
-  def upgradedBuild(logger: Logger, upgrades: Map[ContextualDep, model.Dep], build: model.Build.FileBacked): model.Build.FileBacked = {
-    val newProjects = build.explodedProjects.map { case (crossName, p) =>
-      val versionCombo = model.VersionCombo.fromExplodedProject(p).orThrowTextWithContext(crossName)
-      val newDeps = p.dependencies.map { dep =>
-        upgrades.get((dep, versionCombo)) match {
-          case Some(newDep) =>
-            logger
-              .withContext("project", crossName)
-              .info(s"${dep.organization.value}:${dep.baseModuleName.value} ${dep.version} => ${newDep.version}")
-            newDep
-          case None => dep
-        }
-      }
-      (crossName, p.copy(dependencies = newDeps))
-    }
-
-    BuildRewrite.withProjects(build, newProjects)
-  }
 
   def fetchAllVersions[K](fileCache: FileCache[Task], repos: List[Repository], allDeps: Map[K, Dependency])(implicit
       ec: ExecutionContext
